+++ packages/dartvel_generator/lib/dartvel_generator.dart
@@ -1,2 +1,3 @@
+export 'src/retired.dart' show dartvelGeneratorRetirementNotice;
 export 'src/route_generator.dart' show routeBuilder;
 export 'src/router_builder.dart' show pageBodyBuilder, routerBuilder;